Remove despawn name checks
When the "Show All" configuration is enabled, it is checked for NPC
spawns, but not correctly checked for despawns, leading to these entries
never being cleared. Anyway because calling `Map#remove()` with an entry
for which no key exists is idempotent it is safe to remove any
despawning NPC from these maps regardless of their name.